+++ configParser/Config.cpp
@@ -1,137 +1,23 @@
-#ifndef CONFIG_HPP
-# define CONFIG_HPP
+#include "Config.hpp"
 
-#include "Config_struct.hpp"
-#include "BaseServer.hpp"
-// using namespace std;
-
-
-/*	Request class
-	author	:kbaek
-	date 	:2022.12.10
-
+/*
+	Config_base class
 */
-using vector_int_type = std::vector<int>;
-std::ostream& operator << (std::ostream& os, const vector_int_type& vect) {
-    for (const auto& i : vect)
-        os << ' ' << i;
-    return os;
-}
-
-using vector_str_type = std::vector<std::string>;
-std::ostream& operator << (std::ostream& os, const vector_str_type& vect) {
-    for (const auto& i : vect)
-        os << ' ' << i;
-    return os;
-}
-
-
-struct Config_base
-{
-	Config_base() { numOfServer = 0;}
-	std::vector<BaseServer> getConfigBase();
-	int	getNumOfServer();
-	void print_config();
-
-protected:
-	std::vector<BaseServer> 		base;
-	int								numOfServer;
-};
-
-void Config_base::print_config()
-{
-	BaseServer baseS;
-	ServerInfo tmpserver;
-	std::map< std::string, Location >	tmplocation;
-	std::map< std::string, CgiConfig >	tmpcgi;
-	std::cout << base.size() <<std::endl;
-	std::vector<BaseServer>::iterator it;
-	for (it = base.begin(); it != base.end(); it++)
-	{
-		tmpserver = it->getBServer();
-		tmplocation = it->getBLocation();
-		tmpcgi = it->getBCgi();
-
-		std::cout << "\n\n*** ServerInfo *** \n";
-		std::cout << "serverName = [" << tmpserver.serverName << "]\nhost= [" << tmpserver.host << "]\nport = [" << tmpserver.port << \
-				"]\nmaxRequestBodySize = [" << tmpserver.maxRequestBodySize << "]\n"<<std::endl;
-		std::cout << "\n** Error page ** \n";
-		for (auto& item : tmpserver.errorPages)
-        	std::cout << "key = [" << item.first << "] value = [" << item.second << "]" <<std::endl;
-		
-		std::cout << "\n** Location ** \n";
-		for (auto& itemL : tmplocation)
-		{
-        	std::cout << "\nkey = [" << itemL.first << "]"<< std::endl;
-			Location tmpL;
-			tmpL = itemL.second;
-			std::cout << "value = \n";
-			std::cout << "	root = [" << tmpL.root << "]\n";
-			std::cout << "	maxBody = [" << tmpL.maxBody << "]\n";
-			std::cout << "	returnType = [" << tmpL.returnType << "]\n";
-			std::cout << "	autoListing = [";
-			if (tmpL.autoListing)
-				std::cout << "on]\n";
-			else
-				std::cout << "off]\n";
-			std::cout << "	returnRoot = [" << tmpL.returnRoot << "]\n";
-			std::cout << "	methods = [" << tmpL.Methods << "]\n";
-			std::cout << "	index = [" << tmpL.index << "]\n";
-		}
-
-		std::cout << "\n** Cgi Config ** \n";
-		for (auto& itemC : tmpcgi)
-		{
-        	std::cout << "\nkey = [" << itemC.first << "]"<< std::endl;
-			CgiConfig tmpC;
-			tmpC = itemC.second;
-			std::cout << "value = \n";
-			std::cout << "	root = [" << tmpC.root << "]\n";
-			std::cout << "	methods = [" << tmpC.Methods << "]\n";
-
-		}
-	}
-
-}
-
-std::vector<BaseServer> Config_base::getConfigBase()
+std::vector<BaseServer>
+Config_base::getConfigBase()
 {
 	return this->base;
 }
-int	Config_base::getNumOfServer()
+
+int
+Config_base::getNumOfServer()
 {
 	return this->numOfServer;
 }
 
-class Config : public Config_base
-{
-public:
-	Config();
-	Config(Config const &src);
-	// Config &operator=(Config const &srcs) {};
-	~Config();
-	Config(const std::string path);
-
-	class	FileNotFoundException: public std::exception{
-		virtual const char	*what() const throw();
-	};
-
-	//int check_bracket(int type, int bracket);
-
-private:
-	void configInit(const std::string path);
-	void configParse();
-	void serverInit(int, int);
-	void printV(std::vector <std::string> const &a);
-	
-
-
-private:
-	std::vector<std::string> file;
-};
-
-std::ostream &operator<<(std::ostream &ost, const Config &conf);
-
+/*
+	Config class
+*/
 Config::Config()
 {}
 
@@ -153,7 +39,8 @@ Config::Config(const std::string _path) : Config_base() {
 	return ;
 }
 
-void	Config::configInit(const std::string _path) {
+void
+Config::configInit(const std::string _path) {
 
 
 	std::string buffer;
@@ -172,7 +59,8 @@ void	Config::configInit(const std::string _path) {
 	// printV(file);
 }
 
-void	Config::configParse()
+void
+Config::configParse()
 {
 	unsigned int	fileSize;
 	int 			flag = 0;
@@ -206,11 +94,11 @@ void	Config::configParse()
 				exit (1);
 		}
 	}
-	print_config();
-
+	// print_config();
 }
 
-void	Config::serverInit(int start, int end)
+void
+Config::serverInit(int start, int end)
 {
 	size_t p;
 	size_t sub;
@@ -227,7 +115,7 @@ void	Config::serverInit(int start, int end)
 				sub = file[start].find(" ") + 1;
 				tmpServer.setBServer("h", file[start].substr(sub, p - sub));
 				std::string port = file[start].substr(p + 1);
-				if (!str_is_digit(port))
+				if (!tmpServer.str_is_digit(port))
 				{
 					std::cerr << "Error: listen port is not digit error" << std::endl;
 					exit (1);
@@ -260,7 +148,7 @@ void	Config::serverInit(int start, int end)
 			std::string temp;
 			while (getline(ss, temp, ' '))
 			{
-				if (!str_is_digit(temp))
+				if (!tmpServer.str_is_digit(temp))
 				{
 					std::cerr << "Error: listen block : port is not digit error" << std::endl;
 					exit (1);
@@ -327,7 +215,7 @@ void	Config::serverInit(int start, int end)
 					else if (tmp == "max_body")
 					{
 						std::string maxB = file[start].substr(sub + 1);
-						if (!str_is_digit(maxB))
+						if (!tmpServer.str_is_digit(maxB))
 						{
 							std::cerr << "Error: max_body block : max_body is not digit error" << std::endl;
 							exit (1);
@@ -363,11 +251,6 @@ void	Config::serverInit(int start, int end)
 					exit (1);
 				}
 			}
-			// if (flag)
-			// {
-			// 	std::cerr << "Error: Location block {  } syntax error" << std::endl;
-			// 	exit (1);
-			// }
 			tmpServer.setBlcation(str, location);
 		}
 		else if (file[start].find("cgi ") != std::string::npos)
@@ -414,11 +297,6 @@ void	Config::serverInit(int start, int end)
 					exit (1);
 				}
 			}
-			// if (flag)
-			// {
-			// 	std::cerr << "Error: Location block {  } syntax error" << std::endl;
-			// 	exit (1);
-			// }
 			tmpServer.setBCgi(str, cgi);
 		}
 		else if (file[start] != "server {" && file[start] != "}")
@@ -439,10 +317,8 @@ void	Config::serverInit(int start, int end)
 	*/
 }
 
-
-
-
-void Config::printV(std::vector <std::string> const &a) {
+void
+Config::printV(std::vector <std::string> const &a) {
 
 
 	for(int i=0; i < a.size(); i++)
@@ -451,8 +327,76 @@ void Config::printV(std::vector <std::string> const &a) {
 }
 
 
-const char	*Config::FileNotFoundException::what() const throw(){
+const char*
+Config::FileNotFoundException::what() const throw(){
 	return "Exception thrown: could not open configuration file";
 }
 
-#endif
+// using vector_int_type = std::vector<int>;
+// std::ostream& operator << (std::ostream& os, const vector_int_type& vect) {
+//     for (const auto& i : vect)
+//         os << ' ' << i;
+//     return os;
+// }
+
+// using vector_str_type = std::vector<std::string>;
+// std::ostream& operator << (std::ostream& os, const vector_str_type& vect) {
+//     for (const auto& i : vect)
+//         os << ' ' << i;
+//     return os;
+// }
+
+// void Config_base::print_config()
+// {
+// 	BaseServer baseS;
+// 	ServerInfo tmpserver;
+// 	std::map< std::string, Location >	tmplocation;
+// 	std::map< std::string, CgiConfig >	tmpcgi;
+// 	std::cout << base.size() <<std::endl;
+// 	std::vector<BaseServer>::iterator it;
+// 	for (it = base.begin(); it != base.end(); it++)
+// 	{
+// 		tmpserver = it->getBServer();
+// 		tmplocation = it->getBLocation();
+// 		tmpcgi = it->getBCgi();
+
+// 		std::cout << "\n\n*** ServerInfo *** \n";
+// 		std::cout << "serverName = [" << tmpserver.serverName << "]\nhost= [" << tmpserver.host << "]\nport = [" << tmpserver.port << \
+// 				"]\nmaxRequestBodySize = [" << tmpserver.maxRequestBodySize << "]\n"<<std::endl;
+// 		std::cout << "\n** Error page ** \n";
+// 		for (auto& item : tmpserver.errorPages)
+//         	std::cout << "key = [" << item.first << "] value = [" << item.second << "]" <<std::endl;
+		
+// 		std::cout << "\n** Location ** \n";
+// 		for (auto& itemL : tmplocation)
+// 		{
+//         	std::cout << "\nkey = [" << itemL.first << "]"<< std::endl;
+// 			Location tmpL;
+// 			tmpL = itemL.second;
+// 			std::cout << "value = \n";
+// 			std::cout << "	root = [" << tmpL.root << "]\n";
+// 			std::cout << "	maxBody = [" << tmpL.maxBody << "]\n";
+// 			std::cout << "	returnType = [" << tmpL.returnType << "]\n";
+// 			std::cout << "	autoListing = [";
+// 			if (tmpL.autoListing)
+// 				std::cout << "on]\n";
+// 			else
+// 				std::cout << "off]\n";
+// 			std::cout << "	returnRoot = [" << tmpL.returnRoot << "]\n";
+// 			std::cout << "	methods = [" << tmpL.Methods << "]\n";
+// 			std::cout << "	index = [" << tmpL.index << "]\n";
+// 		}
+
+// 		std::cout << "\n** Cgi Config ** \n";
+// 		for (auto& itemC : tmpcgi)
+// 		{
+//         	std::cout << "\nkey = [" << itemC.first << "]"<< std::endl;
+// 			CgiConfig tmpC;
+// 			tmpC = itemC.second;
+// 			std::cout << "value = \n";
+// 			std::cout << "	root = [" << tmpC.root << "]\n";
+// 			std::cout << "	methods = [" << tmpC.Methods << "]\n";
+
+// 		}
+// 	}
+// }
